requirements.txt
aiohttp==3.6.2
pillow==7.2.0
